[EOC-39][IMP]Improvement code in webhook. [IMP]refactor helpers functions to runbot_helpers.py [IMP]Adds delete xml for old cron [IMP]Update runbot_webhooks.xml [IMP]Adds new condictions for the stickies list into update_git function.
