+++ runbot_webhooks/models/runbot_repo.py
@@ -1,143 +1,49 @@
+# -*- encoding: utf-8 -*-
+##############################################################################
+#
+#    Odoo, Open Source Management Solution
+#    Copyright (C) 2010-2015 Eezee-It (<http://www.eezee-it.com>).
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
 import datetime
-import fcntl
-import glob
-import hashlib
-import itertools
-import logging
-import operator
 import os
-import re
-import resource
-import shutil
-import signal
-import simplejson
-import socket
-import subprocess
-import sys
-import time
-from collections import OrderedDict
 import dateutil.parser
 import logging
 
+# Importing helper functions
+from .runbot_helpers import run
+from .runbot_helpers import decode_utf
 
 from openerp.models import Model, api
 from openerp.fields import Char, Boolean
 
-
-
 _logger = logging.getLogger(__name__)
 _logger.setLevel(logging.DEBUG)
-
-
-#----------------------------------------------------------
-# RunBot helpers
-#----------------------------------------------------------
-
-def log(*l, **kw):
-    out = [i if isinstance(i, basestring) else repr(i) for i in l] + \
-          ["%s=%r" % (k, v) for k, v in kw.items()]
-    _logger.debug(' '.join(out))
-
-def dashes(string):
-    """Sanitize the input string"""
-    for i in '~":\'':
-        string = string.replace(i, "")
-    for i in '/_. ':
-        string = string.replace(i, "-")
-    return string
-
-def mkdirs(dirs):
-    for d in dirs:
-        if not os.path.exists(d):
-            os.makedirs(d)
-
-def grep(filename, string):
-    if os.path.isfile(filename):
-        return open(filename).read().find(string) != -1
-    return False
-
-def rfind(filename, pattern):
-    """Determine in something in filename matches the pattern"""
-    if os.path.isfile(filename):
-        regexp = re.compile(pattern, re.M)
-        with open(filename, 'r') as f:
-            if regexp.findall(f.read()):
-                return True
-    return False
-
-def lock(filename):
-    fd = os.open(filename, os.O_CREAT | os.O_RDWR, 0600)
-    fcntl.lockf(fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
-
-def locked(filename):
-    result = False
-    try:
-        fd = os.open(filename, os.O_CREAT | os.O_RDWR, 0600)
-        try:
-            fcntl.lockf(fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
-        except IOError:
-            result = True
-        os.close(fd)
-    except OSError:
-        result = False
-    return result
-
-def nowait():
-    signal.signal(signal.SIGCHLD, signal.SIG_IGN)
-
-def run(l, env=None):
-    """Run a command described by l in environment env"""
-    log("run", l)
-    env = dict(os.environ, **env) if env else None
-    if isinstance(l, list):
-        if env:
-            rc = os.spawnvpe(os.P_WAIT, l[0], l, env)
-        else:
-            rc = os.spawnvp(os.P_WAIT, l[0], l)
-    elif isinstance(l, str):
-        tmp = ['sh', '-c', l]
-        if env:
-            rc = os.spawnvpe(os.P_WAIT, tmp[0], tmp, env)
-        else:
-            rc = os.spawnvp(os.P_WAIT, tmp[0], tmp)
-    log("run", rc=rc)
-    return rc
-
-def now():
-    return time.strftime(openerp.tools.DEFAULT_SERVER_DATETIME_FORMAT)
-
-def dt2time(datetime):
-    """Convert datetime to time"""
-    return time.mktime(time.strptime(datetime, openerp.tools.DEFAULT_SERVER_DATETIME_FORMAT))
-
-def s2human(time):
-    """Convert a time in second into an human readable string"""
-    for delay, desc in [(86400,'d'),(3600,'h'),(60,'m')]:
-        if time >= delay:
-            return str(int(time / delay)) + desc
-    return str(int(time)) + "s"
-
-def flatten(list_of_lists):
-    return list(itertools.chain.from_iterable(list_of_lists))
-
-def decode_utf(field):
-    try:
-        return field.decode('utf-8')
-    except UnicodeDecodeError:
-        return ''
-
-def uniq_list(l):
-    return OrderedDict.fromkeys(l).keys()
-
-def fqdn():
-    return socket.getfqdn()
 
 class RunbotRepo(Model):
     _inherit = "runbot.repo"
 
-    stickies = Char(string="Stickies", help="Comma-separated list of branch to pull.")
-    auto = Boolean(string="Auto", default=False)
+    # Stickies field for specify the branch to download
+    stickies = Char(string="Stickies", required=True, default='master',
+        help="Comma-separated list of branch to pull.")
+    auto = Boolean(string="Auto", default=True)
 
+    # Overwrite function in file: /runbot/runbot.py#L256
+    # Adds context variable propagation
     def update(self, cr, uid, ids, context=None):
         for repo in self.browse(cr, uid, ids, context=context):
             self.update_git(cr, uid, repo, context=context)
@@ -145,19 +51,32 @@ class RunbotRepo(Model):
     # Overwrite function in file: /runbot/runbot.py#L260
     # Only load the repositories in the stickies field.
     def update_git(self, cr, uid, repo, context=None):
-        _logger.debug('repo %s updating branches', repo.name)
+        '''
+        This function update and create new branch and repositories
+        :param cr: Database cursor
+        :param uid: User ID
+        :param repo: A browse record of runbot.repo
+        :param context: Odoo context dictionary
+        :return: Nothing
+        '''
+        _logger.info('repo %s updating branches', repo.name)
 
         Build = self.pool['runbot.build']
         Branch = self.pool['runbot.branch']
 
+        #Creating a list object for the stickies values
         stickies_list = repo.stickies.split(',')
 
-        if context.get('prbranch'):
-            stickies_list.append(context['prbranch'])
+        #This variable arrive from the webhook
+        #See file: /runbot/controllers/runbot_webhook.py#L61
+        if context.get('pr_branch'):
+            stickies_list.append(context['pr_branch'])
 
+        #Format the list with the correct structure
         stickies_format_list = ['refs/heads/'+x for x in stickies_list]
 
-        _logger.debug("Direccion del repo %s", repo.path)
+        _logger.debug("Repo path %s", repo.path)
+
         if not os.path.isdir(os.path.join(repo.path)):
             os.makedirs(repo.path)
         if not os.path.isdir(os.path.join(repo.path, 'refs')):
@@ -165,9 +84,10 @@ class RunbotRepo(Model):
         else:
             repo.git(['gc', '--auto', '--prune=all'])
 
-            for x in stickies_list:
-                repo.git(['fetch', '-p', 'origin', '+refs/heads/%s:refs/heads/%s' % (x, x)])
-                repo.git(['fetch', '-p', 'origin', '+refs/pull/*/head:refs/pull/*'])
+        #For each element in the stickies list do a fecht
+        for x in stickies_list:
+            repo.git(['fetch', '-p', 'origin', '+refs/heads/%s:refs/heads/%s' % (x, x)])
+            repo.git(['fetch', '-p', 'origin', '+refs/pull/*/head:refs/pull/*'])
 
         fields = ['refname','objectname','committerdate:iso8601','authorname','authoremail','subject','committername','committeremail']
         fmt = "%00".join(["%("+field+")" for field in fields])
@@ -178,46 +98,50 @@ class RunbotRepo(Model):
 
         for name, sha, date, author, author_email, subject, committer, committer_email in refs:
             # create or get branch
+            branch_id = None
             branch_ids = Branch.search(cr, uid, [('repo_id', '=', repo.id), ('name', '=', name)])
-            _logger.debug("Branch Name: %s", name)
-            _logger.debug('Branch ids %s founds', branch_ids)
+            _logger.info("Branch Name: %s", name)
+            _logger.info('Branch ids %s founds', branch_ids)
             if branch_ids:
                 branch_id = branch_ids[0]
             else:
-                _logger.debug('repo %s found new branch %s', repo.name, name)
+                _logger.info('repo %s found new branch %s', repo.name, name)
                 if name in stickies_format_list:
-                    _logger.debug("Into de Condition before the create")
+                    _logger.info("Creating branch %s", name)
                     branch_id = Branch.create(cr, uid, {'repo_id': repo.id, 'name': name})
-                    _logger.debug("After the create")
-            branch = Branch.browse(cr, uid, [branch_id], context=context)[0]
-            # skip build for old branches
-            if dateutil.parser.parse(date[:19]) + datetime.timedelta(30) < datetime.datetime.now():
-                continue
-            # create build (and mark previous builds as skipped) if not found
-            build_ids = Build.search(cr, uid, [('branch_id', '=', branch.id), ('name', '=', sha)])
-            _logger.debug("THIS IS SHA %s", sha)
-            if not build_ids and name in stickies_format_list:
-                _logger.debug('repo %s branch %s new build found revno %s', branch.repo_id.name, branch.name, sha)
-                build_info = {
-                    'branch_id': branch.id,
-                    'name': sha,
-                    'author': author,
-                    'author_email': author_email,
-                    'committer': committer,
-                    'committer_email': committer_email,
-                    'subject': subject,
-                    'date': dateutil.parser.parse(date[:19]),
-                }
+                    _logger.info("Branch %s created succefull", name)
 
-                if not branch.sticky:
-                    skipped_build_sequences = Build.search_read(cr, uid, [('branch_id', '=', branch.id), ('state', '=', 'pending')],
-                                                                fields=['sequence'], order='sequence asc', context=context)
-                    if skipped_build_sequences:
-                        to_be_skipped_ids = [build['id'] for build in skipped_build_sequences]
-                        Build.skip(cr, uid, to_be_skipped_ids, context=context)
-                        # new order keeps lowest skipped sequence
-                        build_info['sequence'] = skipped_build_sequences[0]['sequence']
-                Build.create(cr, uid, build_info)
+            if branch_id is not None:
+                branch = Branch.browse(cr, uid, [branch_id], context=context)[0]
+                # skip build for old branches
+                if dateutil.parser.parse(date[:19]) + datetime.timedelta(30) < datetime.datetime.now():
+                    continue
+                # create build (and mark previous builds as skipped) if not found
+                build_ids = Build.search(cr, uid, [('branch_id', '=', branch.id), ('name', '=', sha)])
+
+                # create the build if name exist into the stickies list
+                if not build_ids and name in stickies_format_list:
+                    _logger.info('repo %s branch %s new build found revno %s', branch.repo_id.name, branch.name, sha)
+                    build_info = {
+                        'branch_id': branch.id,
+                        'name': sha,
+                        'author': author,
+                        'author_email': author_email,
+                        'committer': committer,
+                        'committer_email': committer_email,
+                        'subject': subject,
+                        'date': dateutil.parser.parse(date[:19]),
+                    }
+
+                    if not branch.sticky:
+                        skipped_build_sequences = Build.search_read(cr, uid, [('branch_id', '=', branch.id), ('state', '=', 'pending')],
+                                                                    fields=['sequence'], order='sequence asc', context=context)
+                        if skipped_build_sequences:
+                            to_be_skipped_ids = [build['id'] for build in skipped_build_sequences]
+                            Build.skip(cr, uid, to_be_skipped_ids, context=context)
+                            # new order keeps lowest skipped sequence
+                            build_info['sequence'] = skipped_build_sequences[0]['sequence']
+                    Build.create(cr, uid, build_info)
 
         # skip old builds (if their sequence number is too low, they will not ever be built)
         skippable_domain = [('repo_id', '=', repo.id), ('state', '=', 'pending')]
@@ -226,6 +150,8 @@ class RunbotRepo(Model):
         to_be_skipped_ids = Build.search(cr, uid, skippable_domain, order='sequence desc', offset=running_max)
         Build.skip(cr, uid, to_be_skipped_ids)
 
+    # Overwrite function in file: /runbot/runbot.py#L401
+    # Adds context variable propagation
     def cron(self, cr, uid, ids=None, context=None):
         if ids is None:
             ids = self.search(cr, uid, [('auto', '=', True)], context=context)
