+++ runbot_webhooks/views/runbot_webhook.xml
@@ -6,22 +6,8 @@
       <field name="model">runbot.repo</field>
       <field name="inherit_id" ref="runbot.view_repo_form"/>
       <field name="arch" type="xml">
-<<<<<<< HEAD:runbot_webhooks/views/runbot_webhook.xml
         <field name="nginx" position="after">
           <field name="stickies" string="Stickies"/>
-=======
-
-        <xpath expr="//group[@string='Params']" position="after">
-          <group string="Gitlab">
-            <field name="uses_gitlab"/>
-            <field attrs="{'invisible': [('uses_gitlab','=',False)]}" name="mr_only"/>
-            <field attrs="{'invisible': [('uses_gitlab','=',False)]}" name="sticky_protected"/>
-          </group>
-        </xpath>
-
-        <field name="token" position="attributes">
-          <attribute name="string">API Token</attribute>
->>>>>>> 8.0:runbot_gitlab/runbot_repo_view.xml
         </field>
       </field>
     </record>
